Basketball game
This is basket ball game With is the help of AR Foundation and some precreate prehabs. Unity provides support api for android and ios for creating application. ARFreeThrow is scene file ball prefab and hoop prefab are 3d model and two script ballcontrol which is control the ball and placehoop script is place the hoop where we place the our hoop on surface.